add code for redux functionalitie in bookreducer
